+++ lib/neoid/search_session.rb
@@ -1,8 +1,8 @@
 module Neoid
   class SearchSession
-    def initialize(response, model)
+    def initialize(response, *models)
       @response = response || []
-      @model = model
+      @models = models
     end
     
     def hits
@@ -10,11 +10,21 @@ module Neoid
         Neography::Node.new(x)
       }
     end
+
+    def ids
+      @response.collect { |x| x['data']['ar_id'] }
+    end
     
     def results
-      ids = @response.collect { |x| x['data']['ar_id'] }
-      
-      @model.where(id: ids)
+      models_by_name = @models.inject({}) { |all, curr| all[curr.name] = curr; all }
+
+      ids_by_klass = @response.inject({}) { |all, curr|
+        klass_name = curr['data']['ar_type']
+        (all[models_by_name[klass_name]] ||= []) << curr['data']['ar_id']
+        all
+      }
+
+      ids_by_klass.map { |klass, ids| klass.where(id: ids) }.flatten
     end
   end
 end