+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.0.4
+
+* rewrote seacrch. one index for all types instead of one for type. please run neo_search_index on all of your models.
+  search in multiple types at once with `Neoid.search(types_array, term)
+
 ## v0.0.3
 
 * new configuration syntax (backwards compatible)